requirements_new.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
asyncssh>=2.13.0
cryptography>=41.0.0

# Web API dependencies
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
websockets>=11.0.0

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.10.0

# SSH and execution
paramiko>=3.0.0
jinja2>=3.1.0

# Additional utilities
python-multipart>=0.0.6
pydantic>=2.0.0
